conditionally patch nvfuser

+++ nvfuser_python_utils/patch_nvfuser.py
@@ -8,10 +8,15 @@ def patch_pytorch_nvfuser_binaries(torch_lib):
     import shutil
 
     for f_name in ["libnvfuser_codegen.so"]:
-        shutil.copyfile(
-            os.path.join(nvfuser_lib, f_name),
-            os.path.join(torch_lib, f_name),
-        )
+        target_lib_path = os.path.join(torch_lib, f_name)
+
+        # Only overwrite target torch/lib/libnvfuser_codegen.so if it exists. (before nvfuser removal from torch @ torch <= 2.1)
+        # Don't copy anything there if it doesn't exist. (after nvfuser removal from torch @ torch > 2.1)
+        if os.path.isfile(target_lib_path):
+            shutil.copyfile(
+                os.path.join(nvfuser_lib, f_name),
+                target_lib_path
+            )
 
 
 def remove_nvfuser_python_module(installed_nvfuser_dir):
